# swagger.json
{"swagger":"2.0","info":{"description":"Api Documentation","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"course-controller","description":"Course Controller"},{"name":"university-controller","description":"University Controller"}],"paths":{"/course":{"post":{"tags":["course-controller"],"summary":"createCourse","operationId":"createCourseUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"course","description":"course","required":true,"schema":{"$ref":"#/definitions/Course"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Course"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/course/{id}":{"get":{"tags":["course-controller"],"summary":"getCourseById","operationId":"getCourseByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Course"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["course-controller"],"summary":"updateCourse","operationId":"updateCourseUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"courseDetails","description":"courseDetails","required":true,"schema":{"$ref":"#/definitions/Course"}},{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Course"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["course-controller"],"summary":"deleteCourse","operationId":"deleteCourseUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/courses":{"get":{"tags":["course-controller"],"summary":"getAllCourses","operationId":"getAllCoursesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Course"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/universities":{"get":{"tags":["university-controller"],"summary":"getAllUniversities","operationId":"getAllUniversitiesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/University"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/university":{"post":{"tags":["university-controller"],"summary":"createUniversity","operationId":"createUniversityUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"university","description":"university","required":true,"schema":{"$ref":"#/definitions/University"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/University"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/university/{id}":{"get":{"tags":["university-controller"],"summary":"getUniversityById","operationId":"getUniversityByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/University"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["university-controller"],"summary":"updateUniversity","operationId":"updateUniversityUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"universityDetails","description":"universityDetails","required":true,"schema":{"$ref":"#/definitions/University"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/University"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["university-controller"],"summary":"deleteUniversity","operationId":"deleteUniversityUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"type":"boolean"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}}},"definitions":{"Contact":{"type":"object","properties":{"email":{"type":"string"},"phone":{"type":"string"}},"title":"Contact"},"Course":{"type":"object","properties":{"description":{"type":"string"},"duration":{"type":"string"},"entryRequirements":{"type":"string"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"university":{"$ref":"#/definitions/University"}},"title":"Course"},"University":{"type":"object","properties":{"address":{"type":"string"},"contact":{"$ref":"#/definitions/Contact"},"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"University"}}}